cc3200: fix telnet login for some telnet clients
Telnet clients which dont send the user and password as a single
buffered line will be broken with the existing implementation.
Instead, buffer chars until either a '\r' or the max length of
the username/password are received.

This is required for compatability with Tera Term, common on
windows clients.