unix: Refactor mp_sys_path initialization
Split initialization phase into actual initialization of the list object
and setting the path to a default or a user-supplied value.
The first assuress uPy is setup completely after calling initialize() and
makes initialization similar to how it is done for mp_sys_argv.
The latter allows configuring the path using defaults or an arbitrary value
which comes in handy when embedding.