py/gc.c: decouple gc_sweep from allocation/free operations
This commit is the first part in a series of commits to decouple
the alloc/dealloc component of the upython memory manager from the
garbage collector.

- mem_free implemented to be independent of gc_free.
- mem_first and mem_next functions for a standard api to walk
    through memory

This commit WILL have a small impact on performance. The mem_next
function is not *quite* as fast as the original implementation.
Also, mem_free performs some checks (in MEM_VALID) that have
essentially already been done by the loop.

I contend the following:
- there is no way to make the "next" funcionality faster and allow
    it to be fully decoupled
- we could make the mem_free function faster by implementing a
    mem_free_unsafe (or something) that would not perform the checks.
    This could be a macro for speed, and called by the mem_free
    function. Whether this complexity is required for the small
    increase in performance is up to the upython developers