py: Make stream readall return when bytes_read < bytes_requested.
Before this patch a readall would only return because of a stream
error or when getting 0 bytes from the stream object. This works
well with files but not with blocking sockets.
Suppose you have 50 bytes received on a socket and you call readall,
the first time it tries to read 256 bytes (default read size), but it
only gets 50, so it tries to read again asking for 206 bytes, but since
there's no more data available, the operation blocks indefinitely,
hopefully until the socket is closed by the other end.