py/compile: Remove "return if" optimisation, it's rarely used.
This code would optimise "return x if y else z" patterns, to save 2 bytes
of RAM in the generated bytecode and eliminate a jump.  It costs 56 bytes
of code.  It was there originally to agree with CPython generated
bytecode.

This coding pattern is not common and there are better places to spend
56 bytes of code space.