stmhal: Add ioctl call for enabling/disabling Control-C on UART REPLs.
tested using this file as boot.py:
https://github.com/dhylands/upy-examples/blob/master/boot-cdc.py
(i.e. sets up a repl_uart on one of the UARTs)

Used this test file:
https://github.com/dhylands/upy-examples/blob/master/test_ctrl_c.py

And verified that in each of the 4 cases, pressing Control-C when
the interrupt_char was set to -1 caused no KeyboardInterrupt to
be generated, and pressing Control-C when the interrupt_char was
set to 3 caused a KeyboardInterrupt to be generated.

I had to add ioctl and IOCTL_SET_INTERRUPT_CHAR to teensy and
cc3200 qstrdefsport.h files since they use the stmhal pybstdio.c