py: Add MICROPY_PERSISTENT_NATIVE option for dynamic loadable modules.
With this option enabled MicroPython supports loading of .mpy files
that contain code compiled directly from C (ie dynamic loadable
modules).  Position independent code is enabled by having 2 "link"
tables: one for runtime functions and constants, and one for qstrs that
are local to the loaded code.  The runtime function table is shared with
that used by the native emitter.  The qstr table needs to be populated
by the loaded module on loading of this module.