unix: Make REPL check for pending exceptions on empty lines
This assures pending KeyboardInterrupts generated by pressing Ctrl-C in the
previously executed code are propagated to the REPL when pressing Enter or Ctrl-C;
else the pending exception will only propagate upon the next execution of code.
This is mainly needed for the windows ports as Ctrl-C gets handled in a seperate
thread and as such the exception might be set only by the time the REPL loop is
again waiting for input.