Make install more compatible in unix/Makefile
The current install command uses the flag -D which is specific to the
install command from GNU coreutils, but isn't available for the BSD
version. This solution uses the -d flag which should be commonly
available to create the target directory. Afterwards the target files
are installed to this directory seperately.

Fixes #1984 (greetings to Orwell ;P)